Split glsl type and webGL conversions into separate files
