+++ lib/processPbrMetallicRoughness.js
@@ -4,6 +4,8 @@ var addToArray = require('./addToArray');
 var ForEach = require('./ForEach');
 var numberOfComponentsForType = require('./numberOfComponentsForType');
 var techniqueParameterForSemantic = require('./techniqueParameterForSemantic');
+var webGLConstantToGlslType = require('./webGLConstantToGlslType');
+var glslTypeToWebGLConstant = require('./glslTypeToWebGLConstant')
 
 var clone = Cesium.clone;
 var defined = Cesium.defined;
@@ -70,50 +72,6 @@ function processPbrMetallicRoughness(gltf, options) {
     }
 
     return gltf;
-}
-
-function webGLConstantToGlslType(webGLValue) {
-    switch (webGLValue) {
-        case WebGLConstants.FLOAT:
-            return 'float';
-        case WebGLConstants.FLOAT_VEC2:
-            return 'vec2';
-        case WebGLConstants.FLOAT_VEC3:
-            return 'vec3';
-        case WebGLConstants.FLOAT_VEC4:
-            return 'vec4';
-        case WebGLConstants.FLOAT_MAT2:
-            return 'mat2';
-        case WebGLConstants.FLOAT_MAT3:
-            return 'mat3';
-        case WebGLConstants.FLOAT_MAT4:
-            return 'mat4';
-        case WebGLConstants.SAMPLER_2D:
-            return 'sampler2D';
-        case WebGLConstants.BOOL:
-            return 'bool';
-    }
-}
-
-function glslTypeToWebGLConstant(glslType) {
-    switch (glslType) {
-        case 'float':
-            return WebGLConstants.FLOAT;
-        case 'vec2':
-            return WebGLConstants.FLOAT_VEC2;
-        case 'vec3':
-            return WebGLConstants.FLOAT_VEC3;
-        case 'vec4':
-            return WebGLConstants.FLOAT_VEC4;
-        case 'mat2':
-            return WebGLConstants.FLOAT_MAT2;
-        case 'mat3':
-            return WebGLConstants.FLOAT_MAT3;
-        case 'mat4':
-            return WebGLConstants.FLOAT_MAT4;
-        case 'sampler2D':
-            return WebGLConstants.SAMPLER_2D;
-    }
 }
 
 function generateTechnique(gltf, material, optimizeForCesium) {
